Add async wrappers
