+++ fitscube/logging.py
@@ -70,3 +70,4 @@ ch = logging.StreamHandler()
 ch.setFormatter(formatter)
 ch.setLevel(logging.WARNING)
 logger.addHandler(ch)
+TQDM_OUT = TqdmToLogger(logger, level=logging.INFO)
